requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-environ==0.12.0
idna==3.11
numpy==2.3.5
pillow==12.0.0
psycopg2-binary==2.9.11
reportlab==4.4.4
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
